Rename package; add link to repo
