+++ android/app/src/main/java/com/github/valnesfjord/tg_ws_proxy_rs/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.github.valnesfjord.tgwsproxy
+package com.github.valnesfjord.tg_ws_proxy_rs
 
 import android.Manifest
 import android.content.pm.PackageManager
@@ -12,8 +12,8 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.compose.runtime.LaunchedEffect
 import androidx.core.content.ContextCompat
-import io.github.valnesfjord.tgwsproxy.ui.ProxyScreen
-import io.github.valnesfjord.tgwsproxy.ui.TgWsTheme
+import com.github.valnesfjord.tg_ws_proxy_rs.ui.ProxyScreen
+import com.github.valnesfjord.tg_ws_proxy_rs.ui.TgWsTheme
 
 class MainActivity : ComponentActivity() {
     private val viewModel: ProxyViewModel by viewModels()
