+++ android/app/src/main/res/drawable/ic_github.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="98dp"
+    android:height="96dp"
+    android:viewportWidth="98"
+    android:viewportHeight="96">
+  <group>
+    <clip-path
+        android:pathData="M0,0h98v96h-98z"/>
+    <path
+        android:pathData="M41.439,69.385C28.807,67.854 19.906,58.762 19.906,46.99C19.906,42.205 21.629,37.037 24.5,33.592C23.256,30.434 23.447,23.734 24.883,20.959C28.711,20.48 33.879,22.49 36.941,25.266C40.578,24.117 44.406,23.543 49.096,23.543C53.785,23.543 57.613,24.117 61.059,25.17C64.025,22.49 69.289,20.48 73.117,20.959C74.457,23.543 74.648,30.242 73.404,33.496C76.467,37.133 78.094,42.014 78.094,46.99C78.094,58.762 69.193,67.662 56.369,69.289C59.623,71.395 61.824,75.988 61.824,81.252L61.824,91.205C61.824,94.076 64.217,95.703 67.088,94.555C84.41,87.951 98,70.629 98,49.191C98,22.107 75.988,0 48.904,0C21.82,0 -0,22.107 -0,49.191C-0,70.438 13.494,88.047 31.678,94.65C34.262,95.607 36.75,93.885 36.75,91.301L36.75,83.645C35.41,84.219 33.688,84.602 32.156,84.602C25.84,84.602 22.107,81.156 19.428,74.744C18.375,72.16 17.227,70.629 15.025,70.342C13.877,70.246 13.494,69.768 13.494,69.193C13.494,68.045 15.408,67.184 17.322,67.184C20.098,67.184 22.49,68.906 24.979,72.447C26.893,75.223 28.902,76.467 31.295,76.467C33.688,76.467 35.219,75.605 37.42,73.404C39.047,71.777 40.291,70.342 41.439,69.385Z"
+        android:fillColor="#000000"/>
+  </group>
+</vector>
