+++ android/app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">TG WS Proxy</string>
+    <string name="app_name">TG WS Proxy Rust</string>
     <string name="args_label">Command-line arguments</string>
     <string name="start">Start</string>
     <string name="stop">Stop</string>
@@ -11,4 +11,5 @@
     <string name="log_label">Log</string>
     <string name="notification_channel">Proxy</string>
     <string name="telegram_missing">Telegram is not installed</string>
+    <string name="view_on_github">View on GitHub</string>
 </resources>
